Add a MeasureType to Quake.
Adds a new type for the result of a measurement operation. A value of
type `!quake.measurement` must be post-processed by the new operation
`quake.discriminate` to convert that value to a classical integer value.
`quake.discriminate`, like the previous measurement ops, can produce
either a single integer (bit) or a `std::vector` of integers (bits).

This splits the measurement operation into two steps:
  - acquiring the measurement on a qubit(s)
  - converting the measurement results to a classical value(s).

Fix all the tests and passes to accept the new representation.